Open and closed positions in the same recyclerview

+++ app/src/main/java/mobile/pk/com/stocktracker/transaction/handler/BuyTransactionHandler.java
@@ -11,7 +11,7 @@ public class BuyTransactionHandler implements TransactionHandler{
 
     @Override
     public void apply(Position position, UserTransaction userTransaction) {
-        double quantity = position.getQuantity() + userTransaction.getQuantity();
+        int quantity = position.getQuantity() + userTransaction.getQuantity();
         double avgPrice = ((position.getAveragePrice() * position.getQuantity())  + (userTransaction.getPrice() * userTransaction.getQuantity()))/ (quantity);
         position.setQuantity(quantity);
         position.setAveragePrice(avgPrice);
